+++ molecularnodes/io/parse/star.py
@@ -1,7 +1,17 @@
 import numpy as np
+import bpy
 from .ensemble import Ensemble
 from ... import blender as bl
 
+@bpy.app.handlers.persistent
+def _rehydrate_ensembles(scene):
+    for obj in bpy.data.objects:
+        if hasattr(obj, 'mn') and 'molecule_type' in obj.mn.keys():
+            if obj.mn['molecule_type'] == 'star':
+                ensemble = StarFile.from_blender_object(obj)
+                if not hasattr(bpy.types.Scene, 'MN_starfile_ensembles'):
+                    bpy.types.Scene.MN_starfile_ensembles = []
+                bpy.types.Scene.MN_starfile_ensembles.append(ensemble)
 
 class StarFile(Ensemble):
     def __init__(self, file_path):
